+++ tests/client.py
@@ -4,5 +4,5 @@ from lapidary.runtime import ClientBase
 
 
 class ClientTestBase(ClientBase):
-    def __init__(self, client: httpx.AsyncClient):
-        super().__init__(session_factory=lambda **_: client)
+    def __init__(self, client: httpx.AsyncClient | None = None):
+        super().__init__(client=client)